Update download button
