+++ app/routes/_index.tsx
@@ -4,10 +4,12 @@ import GamePage from "~/components/GamePage";
 import { LinksFunction } from "@remix-run/node";
 import globalStyle from "~/styles/global.css?url";
 import typingBoxStyle from "~/styles/typingBox.css?url";
+import downloadBoxStyle from "~/styles/downloadBox.css?url";
 
 export const links: LinksFunction = () => [
   { rel: "stylesheet", href: globalStyle },
   { rel: "stylesheet", href: typingBoxStyle },
+  { rel: "stylesheet", href: downloadBoxStyle },
 ];
 
 export const meta: MetaFunction = () => {
