+++ app/styles/downloadBox.css
@@ -1,0 +1,27 @@
+.game-page > .download-box {
+  position: fixed;
+  right: 0;
+  bottom: 0;
+  padding: 20px;
+
+  display: flex;
+  flex-direction: row;
+}
+
+.game-page > .download-box > button {
+  width: 40px;
+  height: 40px;
+  font-size: 30px;
+  border: none;
+  background-color: transparent;
+
+  display: flex;
+  justify-content: center;
+  align-items: center;
+
+  color: gray;
+  &:hover {
+    color: yellowgreen;
+    font-weight: bold;
+  }
+}
